+++ ecoz2_test.py
@@ -2,11 +2,10 @@
 # export PYTHONPATH=.:ecoz2py
 # python ecoz2_test.py
 #
-from ecoz2py import ecoz2_get_version
-from ecoz2py import ecoz2_prd_show_file
+import ecoz2py as ecoz2
 
-print("VERSION: {}".format(ecoz2_get_version()))
+print("ECOZ2 C version: {}".format(ecoz2.get_version()))
 
-ecoz2_prd_show_file(
+ecoz2.prd_show_file(
   b"../ecoz2-whale/exerc01/data/predictors/TRAIN/B/from_MARS_20161221_000046_SongSession_16kHz_HPF5Hz.wav__10079.092_10080.35.prd"
 )