+++ ecoz2py/ecoz2_extension_test.py
@@ -1,6 +1,6 @@
 from _ecoz2_extension import ffi
 from _ecoz2_extension.lib import ecoz2_version, ecoz2_baz, ecoz2_hi
-from _ecoz2_extension.lib import prd_show_file
+from _ecoz2_extension.lib import ecoz2_prd_show_file
 from _ecoz2_extension.lib import ecoz2_do_filenames
 
 print("ECOZ2 C version: {}".format(ffi.string(ecoz2_version())))
@@ -27,7 +27,7 @@ def myfunc(c_filename, i):
 
 ecoz2_do_filenames(buffer_in, len(filenames), myfunc)
 
-prd_show_file(
+ecoz2_prd_show_file(
   b"../../ecoz2-whale/exerc01/data/predictors/TRAIN/B/from_MARS_20161221_000046_SongSession_16kHz_HPF5Hz.wav__10079.092_10080.35.prd",
   0, -1, -1
 )